@@ -9,7 +9,7 @@ describe RoutesController, :type => :controller do
 
   shared_examples_for "redirected to referential_line_path(referential,line)" do
     it "should redirect_to referential_line_path(referential,line)" do
-      #response.should redirect_to( referential_line_path(referential,route.line) )
+      # expect(response).to redirect_to( referential_line_path(referential,route.line) )
     end
   end
   shared_examples_for "line and referential linked" do
