@@ -6,7 +6,7 @@ require 'spec_helper'
 # describe ExportsHelper do
 #   describe "string concat" do
 #     it "concats two strings with spaces" do
-#       helper.concat_strings("this","that").should == "this that"
+#       expect(helper.concat_strings("this","that")).to == "this that"
 #     end
 #   end
 # end
