@@ -1,7 +1,22 @@
 class SetDefaultValueForDataFormatInOrganisation < ActiveRecord::Migration
   def change
-    Organisation.where(data_format: nil).update_all(data_format: "neptune")
-    execute "update referentials set data_format = organisations.data_format from organisations where referentials.data_format is null and referentials.organisation_id = organisations.id"
+    Organisation.all.each do |organisation|
+      if organisation.data_format.neptune?
+        organisation.update_attribute(:data_format, "neptune")
+      end
+    end
+    Referential.all.each do |referential|
+      if referential.data_format.neptune?
+        referential.update_attribute :data_format, "neptune"
+      elsif referential.data_format.netex?
+        referential.update_attribute :data_format, "netex"
+      elsif referential.data_format.gtfs?
+        referential.update_attribute :data_format, "gtfs"
+      elsif referential.data_format.hub?
+        referential.update_attribute :data_format, "hub"
+      end
+    end
+
     change_column :organisations, :data_format, :string, :default => "neptune"
   end
 end
